Move parse document methods from Document to utils
